LayoutTests/platform/ios/imported/w3c/web-platform-tests/selection/Selection-direction-on-clicks.tentative-expected.txt

hello, world

FAIL direction returns "none" when there is a double click selection(directionless) assert_equals: expected Text node "hello, world" but got null
FAIL direction returns "none" when there is a triple click selection(directionless) assert_equals: expected Element node <div id="container">hello, world</div> but got null
FAIL direction returns "none" when the selection is collapsed assert_equals: expected Text node "hello, world" but got null
